Consolidate metrics updates into one batch job
