+++ functions/src/index.ts
@@ -16,8 +16,7 @@ export * from './on-fold-follow'
 export * from './on-fold-delete'
 export * from './on-view'
 export * from './unsubscribe'
-export * from './update-contract-metrics'
-export * from './update-user-metrics'
+export * from './update-metrics'
 export * from './update-recommendations'
 export * from './backup-db'
 export * from './change-user-info'
