+++ functions/src/scripts/update-metrics.ts
@@ -2,15 +2,12 @@ import { initAdmin } from './script-init'
 initAdmin()
 
 import { log, logMemory } from '../utils'
-import { updateContractMetricsCore } from '../update-contract-metrics'
-import { updateUserMetricsCore } from '../update-user-metrics'
+import { updateMetricsCore } from '../update-metrics'
 
 async function updateMetrics() {
   logMemory()
-  log('Updating contract metrics...')
-  await updateContractMetricsCore()
-  log('Updating user metrics...')
-  await updateUserMetricsCore()
+  log('Updating metrics...')
+  await updateMetricsCore()
 }
 
 if (require.main === module) {
